<wip: documentation for base module and switching to javapoet for code generation at compile time>

+++ base/src/main/java/it/auties/protobuf/annotation/ProtobufReservedRange.java
@@ -1,6 +1,22 @@
 package it.auties.protobuf.annotation;
 
+/**
+ * This annotation represents a range of numeric values that are reserved and cannot be used
+ * in the context of Protobuf definitions. A reserved range ensures compatibility
+ * or avoids conflicts when defining numeric values in Protobuf.
+ */
 public @interface ProtobufReservedRange {
-    int min();
-    int max();
+    /**
+     * Returns the minimum value of the reserved numeric range, inclusive.
+     *
+     * @return the minimum value of the reserved numeric range
+     */
+    long min();
+
+    /**
+     * Returns the maximum value of the reserved numeric range, inclusive.
+     *
+     * @return the maximum value of the reserved numeric range
+     */
+    long max();
 }
